Convert all javatari save objects to use string field names
The Javatari save objects were using object parameters as named parameters instead of strings.  This means that advanced compilation could potentially rename the fields, breaking compatibility of old save objects between builds.  This commit changes all the save objects to use strings for the parameters so that they are guaranteed to remain consistent.