Fix drift issues
While other commits attempted to fix drift issues, and no doubt solved some problems, this commit is definitely getting at a major core reason why they still occurred.  The problem was a common mistake, and one that I should have noticed.  I had set it to perform a reset while in the process of running through some CPU cycles.  Knowing that could be bad, I canceled out the current cycle when resetting, but that didn't take into account that it wasn't running through just one cycle.

So, Javatari wakes up and is going to run through N cycles.  Halfway through, the agent decides to do a reset.   The problem is, Javatari keeps running through the second half of the cycles, making the state we are trying to start at somewhat off.

This solution attempts to use the pause functionality to schedule a pause, which is detected outside of the running through the tics.  When the pause occurs, it is safe to restart, and then safe to go again.   This should (hopefully) make all methods of loading the game work from a consistent point.