Adjust algorithm to train faster (shortRight)
This adds in the concept of shortRight.  When the player is stopped and going to move right, a chanceOfShortRight is checked, and if it occurs they the right is quickly followed by another command.  This makes moving from stopped to jump-right much more likely.

The chanceOfRight was increased by .05 to make the training idle a little less.