Deal with situations where we enter screen on a log
Every once in a while harry will jump into a new screen and land right
on a rolling log.  When he entered the new screen, we checkpointed,
and when he hit the log we reset.  This creates an infinite loop of
reloading the checkpoint, hitting the log, and then once again
reloading the checkpoint.

While it isn't the nicest way, we deal with this by removing the
current checkpoint and all history up until the prior checkpoint.  We
then get rid of our checkpoint state and restart.  This has the
unfortunate effect of sending Harry all the way back to the beginning
to retrace his steps.  He then has to play through the previously
solved last screen and hope to solve and cross over again at a safer
time.